+++ packages/components/atoms/f-image-tile/CHANGELOG.md
@@ -10,6 +10,9 @@ v0.13.0
 ### Added
 - Breakout image option and tests
 
+### Removed
+- unneeded `load` and `waitForComponent` functions from component object
+
 ### Changed
 - Refactor WebDriverIO tests to use async in order to support Node 16 using `codemod` utility.